Extract admin-only AddressType
