+++ src/Sylius/Bundle/AdminBundle/Form/Type/OrderType.php
@@ -15,12 +15,28 @@ namespace Sylius\Bundle\AdminBundle\Form\Type;
 
 use Sylius\Bundle\OrderBundle\Form\Type\OrderType as BaseOrderType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 final class OrderType extends AbstractType
 {
-    public function getBlockPrefix(): string
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        return 'sylius_admin_order';
+        $builder
+            ->add('shippingAddress', AddressType::class, [
+                'add_build_address_form_subscriber' => $options['add_build_address_form_subscriber'],
+            ])
+            ->add('billingAddress', AddressType::class, [
+                'add_build_address_form_subscriber' => $options['add_build_address_form_subscriber'],
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        parent::configureOptions($resolver);
+
+        $resolver->setDefaults(['add_build_address_form_subscriber' => true]);
     }
 
     public function getParent(): string
